requirements.txt
Flask
PyPDF2
reportLab
Pillow
qrcode
docx2pdf
pdf2image
Werkzeug